Helmet Heroes Bot
This is used to see the data in Helmet Heroes accs and other details which people need to know on other users or on themselves. Rn this is just to let me understand how the bot works and focusing on making my own version as well give me ideas to create things.